API: Move DeviceTimePlot to tools subdirectory
