+++ typhon/tools/__init__.py
@@ -1,0 +1,4 @@
+"""Module for all insertable Typhon tools"""
+__all__ = ['DeviceTimePlot']
+
+from .plot import DeviceTimePlot
